Refactor interface pins to be attached to chips
Change so instead of the external pin part of the interface pin attaching to the chip they instead attach to the interface pin and the interface pin attaches to the chip. This simplifies certain parts of the code.

Remove pin, and compute working after some preliminary testing. Icons representations are broken.